refactor: Move the tokio module to its own file
