+++ src/aio/tokio.rs
@@ -1,0 +1,139 @@
+use super::{async_trait, AsyncStream, RedisResult, RedisRuntime, SocketAddr};
+
+use std::{
+    future::Future,
+    io,
+    pin::Pin,
+    task::{self, Poll},
+};
+
+#[cfg(unix)]
+use tokio::net::UnixStream as UnixStreamTokio;
+
+use tokio::{
+    io::{AsyncRead, AsyncWrite},
+    net::TcpStream as TcpStreamTokio,
+};
+
+#[cfg(feature = "tls")]
+use super::TlsConnector;
+
+#[cfg(feature = "tokio-tls-comp")]
+use tokio_tls::TlsStream;
+
+#[cfg(unix)]
+use super::Path;
+
+pub(crate) enum Tokio {
+    /// Represents a Tokio TCP connection.
+    Tcp(TcpStreamTokio),
+    /// Represents a Tokio TLS encrypted TCP connection
+    #[cfg(feature = "tokio-tls-comp")]
+    TcpTls(TlsStream<TcpStreamTokio>),
+    /// Represents a Tokio Unix connection.
+    #[cfg(unix)]
+    Unix(UnixStreamTokio),
+}
+
+impl AsyncWrite for Tokio {
+    fn poll_write(
+        mut self: Pin<&mut Self>,
+        cx: &mut task::Context,
+        buf: &[u8],
+    ) -> Poll<io::Result<usize>> {
+        match &mut *self {
+            Tokio::Tcp(r) => Pin::new(r).poll_write(cx, buf),
+            #[cfg(feature = "tokio-tls-comp")]
+            Tokio::TcpTls(r) => Pin::new(r).poll_write(cx, buf),
+            #[cfg(unix)]
+            #[cfg(feature = "tokio-comp")]
+            Tokio::Unix(r) => Pin::new(r).poll_write(cx, buf),
+        }
+    }
+
+    fn poll_flush(mut self: Pin<&mut Self>, cx: &mut task::Context) -> Poll<io::Result<()>> {
+        match &mut *self {
+            Tokio::Tcp(r) => Pin::new(r).poll_flush(cx),
+            #[cfg(feature = "tokio-tls-comp")]
+            Tokio::TcpTls(r) => Pin::new(r).poll_flush(cx),
+            #[cfg(unix)]
+            Tokio::Unix(r) => Pin::new(r).poll_flush(cx),
+        }
+    }
+
+    fn poll_shutdown(mut self: Pin<&mut Self>, cx: &mut task::Context) -> Poll<io::Result<()>> {
+        match &mut *self {
+            Tokio::Tcp(r) => Pin::new(r).poll_shutdown(cx),
+            #[cfg(feature = "tokio-tls-comp")]
+            Tokio::TcpTls(r) => Pin::new(r).poll_shutdown(cx),
+            #[cfg(unix)]
+            Tokio::Unix(r) => Pin::new(r).poll_shutdown(cx),
+        }
+    }
+}
+
+impl AsyncRead for Tokio {
+    fn poll_read(
+        mut self: Pin<&mut Self>,
+        cx: &mut task::Context,
+        buf: &mut [u8],
+    ) -> Poll<io::Result<usize>> {
+        match &mut *self {
+            Tokio::Tcp(r) => Pin::new(r).poll_read(cx, buf),
+            #[cfg(feature = "tokio-tls-comp")]
+            Tokio::TcpTls(r) => Pin::new(r).poll_read(cx, buf),
+            #[cfg(unix)]
+            Tokio::Unix(r) => Pin::new(r).poll_read(cx, buf),
+        }
+    }
+}
+
+#[async_trait]
+impl RedisRuntime for Tokio {
+    async fn connect_tcp(socket_addr: SocketAddr) -> RedisResult<Self> {
+        Ok(TcpStreamTokio::connect(&socket_addr)
+            .await
+            .map(Tokio::Tcp)?)
+    }
+
+    #[cfg(feature = "tls")]
+    async fn connect_tcp_tls(
+        hostname: &str,
+        socket_addr: SocketAddr,
+        insecure: bool,
+    ) -> RedisResult<Self> {
+        let tls_connector: tokio_tls::TlsConnector = if insecure {
+            TlsConnector::builder()
+                .danger_accept_invalid_certs(true)
+                .danger_accept_invalid_hostnames(true)
+                .use_sni(false)
+                .build()?
+        } else {
+            TlsConnector::new()?
+        }
+        .into();
+        Ok(tls_connector
+            .connect(hostname, TcpStreamTokio::connect(&socket_addr).await?)
+            .await
+            .map(Tokio::TcpTls)?)
+    }
+
+    #[cfg(unix)]
+    async fn connect_unix(path: &Path) -> RedisResult<Self> {
+        Ok(UnixStreamTokio::connect(path).await.map(Tokio::Unix)?)
+    }
+
+    fn spawn(f: impl Future<Output = ()> + Send + 'static) {
+        tokio::spawn(f);
+    }
+
+    fn boxed(self) -> Pin<Box<dyn AsyncStream + Send + Sync>> {
+        match self {
+            Tokio::Tcp(x) => Box::pin(x),
+            #[cfg(feature = "tokio-tls-comp")]
+            Tokio::TcpTls(x) => Box::pin(x),
+            #[cfg(unix)]
+            Tokio::Unix(x) => Box::pin(x),
+        }
+    }
+}
